[ONB-71] Mail 문의 구현  (#7)
* ONB-71 feat: 문의, 지원용 메일 string 추가

* ONB-71 feat: quit 지원용 메일 string으로 변경

* ONB-71 feat: 문의 구글 폼에서 메일로 변경

* ONB-71 style: ktlint formatting

* ONB-71 feat: change support mail, content string

* ONB-71 refactor: mail 전송 유틸 함수로 리팩토링
