+++ presentation/src/main/res/values/strings.xml
@@ -102,4 +102,16 @@
 - 계정 정보 확인이 완료되면 즉시 회원 탈퇴 처리가 완료되어 회원님의 계정 정보와 개인정보가 파기됩니다. 파기된 정보는 복구되지 않습니다. (서비스 제공에 필요한 일부 게임 결과 기록은 식별 불가능하게 처리되어 보존될 수 있습니다.)\n
 - 계정 정보 확인이 어려울 경우 회원 탈퇴가 정상적으로 이루어지지 않습니다.\n
 - 처리 결과를 신청하신 메일로 안내해드립니다. 수신 가능한 메일 주소를 사용해주세요.\n</string>
+    <string name="help_email_content">온보드 지원팀에 문의해주셔서 감사합니다.\n
+사용하신 메일로 처리 결과를 안내해 드립니다. 수신 가능한 메일 주소를 사용해주세요.
+\n\n
+- (자동 입력) 유저 ID : %d
+\n
+- (자동 입력) 닉네임 : %s
+\n
+- (자동 입력) 그룹 ID : %d
+\n
+- 문의 내용 :
+    </string>
+    <string name="onboard_mail">help.onboardgame@gmail.com</string>
 </resources>
